docs: convert playwright-fixtures example app to nitro and demonstrate backend log capture
The app under test is now a small standalone Nitro app (same pages and
endpoints as the plain-node server it replaces) instrumented with
@piwitests/instrumentation via the documented one-line plugin re-export.
A new backend-logs spec drives a consola warning+error route and an
unhandled-500 route, so the dashboard's Network & backend logs panel has
a live demo. Reporter and instrumentation pins move to exact released
versions (^0.11.0 excluded the published 0.12.0), and the suite keeps
its single intentional failure.
