+++ examples/playwright-fixtures/README.md
@@ -2,7 +2,7 @@
 
 A small, runnable Playwright project wired to [Piwi Dashboard](https://piwitests.github.io) with the **[capture fixtures](https://piwitests.github.io/capture-fixtures)** — the one-file addition that unlocks slow-endpoint analysis, Web Vitals, console capture, failure-time ARIA snapshots, and locator healing.
 
-It tests a tiny dependency-free web app (`app/server.mjs`, started automatically by Playwright) and exercises **every capture path** the fixtures support.
+It tests a tiny [Nitro](https://nitro.build) web app (`app/`, started automatically by Playwright) and exercises **every capture path** the fixtures support — including **[backend logs](https://piwitests.github.io/backend-logs)**: the app is instrumented with [`@piwitests/instrumentation`](../../integrations/nitro), so server-side warnings and errors ride back to the dashboard on the `X-Piwi-Logs` response header.
 
 ## Run it
 
@@ -13,6 +13,8 @@ npm install
 npx playwright install chromium
 npm test
 ```
+
+> On a fresh Linux machine (e.g. a CI runner), use `npx playwright install --with-deps chromium` so the browser's system libraries are installed too.
 
 Results appear at `http://localhost:3000` under the `playwright-fixtures-example` project. Point elsewhere with `PIWI_DASHBOARD_URL` / `PIWI_PROJECT_NAME`.
 
@@ -31,6 +33,7 @@ Results appear at `http://localhost:3000` under the `playwright-fixtures-example
 | `failing-locator.spec.ts` | **Intentional failure** → ARIA snapshot, locator suggestion, locator healing |
 | `console.spec.ts` | `console.warn` / `error` / `assert` captured; `console.log` intentionally not |
 | `slow-endpoint.spec.ts` | A slow API call plus `/api/users/1` and `/api/users/2` — grouped as `/api/users/:id` in the *Slow endpoints* tab |
+| `backend-logs.spec.ts` | **Backend logs** — a warning + error logged via `consola`, and an unhandled 500, attached to their network requests via `X-Piwi-Logs` |
 | `skipped.spec.ts` | Skipped tests produce no capture attachments |
 | `before-all.spec.ts` | `beforeAll` activity is intentionally **not** captured |
 | `custom-fixtures.spec.ts` | Dashboard capture composed with your own fixtures (`fixtures-composed.ts`) |
@@ -46,3 +49,18 @@ Every spec imports `test` from one of these files — never from `@playwright/te
 
 - `playwright.config.ts` — **recommended**: `wrapConfig` injects the reporter + global setup (`npm test`)
 - `playwright.manual-reporter.config.ts` — plain `reporter` array (`npm run test:manual-reporter`)
+
+## Backend logs in one file
+
+The entire backend-log setup is `app/plugins/piwi-test-logs.ts`:
+
+```ts
+export { default } from '@piwitests/instrumentation';
+```
+
+Nitro auto-loads it from `plugins/` (in a Nuxt app, put the same file in `server/plugins/`). Two things worth knowing:
+
+- Only **`consola`** warnings/errors and **unhandled request errors** are captured — bare `console.warn` calls are not (see `app/api/report.get.ts` and `app/api/failing.get.ts` for both flavors).
+- Capture is on outside production; set `PIWI_TEST_LOGS_DISABLED=true` to turn it off anywhere, or `PIWI_TEST_LOGS_DISABLED=false` to force it on in a production-mode test deployment.
+
+To poke at it without the dashboard: `npm start`, then open `http://localhost:4173/backend` and watch the `X-Piwi-Logs` response header on `/api/report`.
