+++ examples/playwright-fixtures/app/api/before-all-marker.get.ts
@@ -1,0 +1,3 @@
+import { defineEventHandler } from 'h3';
+
+export default defineEventHandler(() => ({ ok: true }));
