+++ examples/playwright-fixtures/app/api/child.get.ts
@@ -1,0 +1,3 @@
+import { defineEventHandler } from 'h3';
+
+export default defineEventHandler(() => ({ ok: true }));
