+++ examples/playwright-fixtures/app/api/items.get.ts
@@ -1,0 +1,3 @@
+import { defineEventHandler } from 'h3';
+
+export default defineEventHandler(() => [{ name: 'Alpha' }, { name: 'Beta' }, { name: 'Gamma' }]);
