+++ examples/playwright-fixtures/app/api/submit.post.ts
@@ -1,0 +1,6 @@
+import { defineEventHandler, readBody } from 'h3';
+
+export default defineEventHandler(async (event) => {
+  await readBody(event);
+  return { ok: true };
+});
