+++ examples/playwright-fixtures/app/api/users/[id].get.ts
@@ -1,0 +1,9 @@
+import { createError, defineEventHandler, getRouterParam } from 'h3';
+
+export default defineEventHandler((event) => {
+  const id = Number(getRouterParam(event, 'id'));
+  if (!Number.isInteger(id)) {
+    throw createError({ statusCode: 404, statusMessage: 'Not found' });
+  }
+  return { id, name: `User ${id}` };
+});
