+++ examples/playwright-fixtures/package.json
@@ -2,17 +2,21 @@
   "name": "piwi-example-playwright-fixtures",
   "private": true,
   "version": "0.0.0",
-  "description": "Runnable example: a small web app + Playwright tests wired to Piwi Dashboard with the capture fixtures",
+  "description": "Runnable example: a small Nitro web app + Playwright tests wired to Piwi Dashboard with the capture fixtures and backend-log instrumentation",
   "type": "module",
   "scripts": {
-    "start": "node app/server.mjs",
+    "start": "nitro dev --port 4173",
     "test": "playwright test",
     "test:manual-reporter": "playwright test --config=playwright.manual-reporter.config.ts",
     "typecheck": "tsc --noEmit"
   },
   "devDependencies": {
-    "@piwitests/reporter": "^0.11.0",
+    "@piwitests/instrumentation": "0.12.0",
+    "@piwitests/reporter": "0.12.0",
     "@playwright/test": "^1.56.1",
+    "consola": "^3.4.2",
+    "h3": "^1.15.11",
+    "nitropack": "^2.13.4",
     "typescript": "^5.9.3"
   }
 }
